Replace `SearchInnerNodeSorted` with runtime dispatch
**Description**
Remove `SearchInnerNodeSorted` trait that was implemented for specific
sizes of `InnerNodeSorted<..., SIZE>` and replace with a runtime check
that dispatches to a function optimized for specific sizes.

Pull out some common range bound validation code into a helper function.

Switch from using `MaybeUninit<u8>` in the `InnerNodeSorted` type to
plain `u8`, and update related code. Remove some unsound code that
blanket converted from `[MaybeUninit<u8>; 16]` to `[u8; 16]` without
confirming that all values were "init".

**Motivation**
The `SearchInnerNodeSorted` trait made it difficult to make the
`impl InnerNode<...> for InnerNodeSorted<..., SIZE>` generic over
`SIZE`, since it was implemented for specific values of `SIZE`.

Helper function is good to reduce code duplication.

The change from `MaybeUninit<u8>` to `u8` was to remove the complexity
of dealing with those uninit values, specifically when using converting
the keys array to SIMD types. I also think that we won't lose much
performance, since the keys array is typically small and we don't gain
much by leaving those values uninit.

**Testing Done**
`./scripts/full-test.sh nightly`

+++ src/raw/representation/inner_node_indirect/index.rs
@@ -137,7 +137,7 @@ mod tests {
     #[test]
     fn convert() {
         use core::convert::TryFrom;
-        let zero = NonMaxIndex::try_from(0 as u8).unwrap();
+        let zero = NonMaxIndex::try_from(0u8).unwrap();
         let zero = u8::from(zero);
         assert_eq!(zero, 0);
